refactor: move pages and components into separate folders
