+++ src/components/FormCard/FormCard.tsx
@@ -1,0 +1,8 @@
+import React, { Component } from 'react';
+import { v4 as uuidv4 } from 'uuid';
+
+export default class FormCard extends Component {
+  render() {
+    return <div>FormCard</div>;
+  }
+}
