+++ src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Outlet } from 'react-router-dom';
-import Header from './Header';
+import Header from 'components/Header/Header';
 
 export default class Layout extends Component {
   render() {
